realtime_app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
ultralytics==8.3.25
opencv-python-headless==4.10.0.84
numpy>=1.23
